Fix: Explicitly cast signed-unsigned conversions
This patch adds two explicit casts to the `put` function of both
`FileBackedStorage` and `InMemoryStorage`, silencing warnings from
Clang, and making it explicit that we have a type boundary between one
interface that provides us with a signed `int`, and another that
provides us with an `unsigned int`.  A cursory examination of both of
these interfaces does not give a clear answer to whether we can (or
should) change their type; instead we make explicit the unsafe
type-conversion behavior here.

+++ src/groups/mqb/mqbs/mqbs_inmemorystorage.cpp
@@ -174,9 +174,10 @@ InMemoryStorage::put(mqbi::StorageMessageAttributes*     attributes,
 {
     // PRECONDITIONS
     BSLS_ASSERT_SAFE(appData);
-    BSLS_ASSERT_SAFE(appData->length() == attributes->appDataLen());
+    BSLS_ASSERT_SAFE(static_cast<unsigned int>(appData->length()) ==
+                     attributes->appDataLen());
 
-    const int    msgSize  = attributes->appDataLen();
+    const int    msgSize  = static_cast<int>(attributes->appDataLen());
     unsigned int refCount = attributes->refCount();
     // Proxies are unaware of the number of apps unlike Replicas.
     // The latter can check for duplicates.
